user/pyproject.toml
[project]
name = "user"
version = "0.1.0"
description = "User microservice"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accessify>=0.3.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
